MTRF/algorithms/requirements.txt
absl-py==0.7.1
aiohttp==3.5.4
asn1crypto==0.24.0
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
awscli==1.16.140
boto3==1.9.130
botocore==1.12.130
cachetools==3.0.0
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorama==0.3.9
cryptography==2.3.1
cycler==0.10.0
Cython==0.29.1
dask==1.0.0
decorator==4.3.0
docutils==0.14
deepdiff==3.3.0
flatbuffers==1.10
flatten-dict==0.0.3.post1
funcsigs==1.0.2
future==0.17.1
gast==0.2.2
gitdb2==2.0.5
GitPython==2.1.11
glfw==1.7.0
google-api-python-client==1.7.5
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-compute-engine==2.8.13
grpcio==1.16.1
gtimer==1.0.0b5
# gym==0.12.1
gym==0.15.6
h5py==2.8.0
httplib2==0.12.0
idna==2.7
ipdb>=0.12
imageio==2.4.1
ipython==7.3.0
jmespath==0.9.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
lockfile==0.12.2
Markdown==3.0.1
matplotlib==3.0.2
more-itertools==4.3.0
# git+https://github.com/hartikainen/mujoco-py.git@435d2143abe04fe4648c6c0c1b848bf1fc06b73b
# git+https://github.com/avisingh599/multiworld.git@19bf319422c0016260166bf64e194552bf2a9e68
mujoco-py<2.1,>=2.0
networkx==2.2
numpy==1.17.3
opencv-python==4.0.0.21
pandas==0.23.4
Pillow==5.3.0
plotly==1.9.6
pluggy==0.9.0
protobuf==3.6.1
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycosat==0.6.3
pycparser==2.19
pygame==1.9.4
pyglet==1.3.2
pyOpenSSL==18.0.0
pyparsing==2.3.0
PySocks==1.6.8
pytest==4.5.0
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==5.1
psutil==5.4.8
ray[rllib,debug]==0.7.1
redis==3.0.1
requests==2.20.1
# robosuite==0.1.0
rsa==3.4.2
# ruamel-yaml==0.15.46
s3transfer==0.2.0
scikit-image==0.14.2
scikit-video==1.1.11
scikit-learn==0.20.1
scipy==1.1.0
git+https://github.com/hartikainen/serializable.git@76516385a3a716ed4a2a9ad877e2d5cbcf18d4e6
setproctitle==1.1.10
six==1.11.0
smmap2==2.0.5
tensorboard==1.15.0
tensorflow-gpu==1.15
tensorflow-probability==0.7.0
termcolor==1.1.0
toolz==0.9.0
uritemplate==3.0.0
urllib3<1.25,>=1.21.1
Werkzeug==0.14.1
